perf(core): load @types/node and Monaco typings registration lazily
main.ts eagerly inlined the whole @types/node tree as raw strings
(import.meta.glob with eager: true) and imported the Monaco chunk at
boot for every user, editor page or not. The typings alone accounted
for 2.4 MB of the 2.65 MB entry chunk (~444 KB gzipped on the critical
path).

Move the glob into a dedicated module that is only imported
dynamically, and register the extra libs from MonacoEnvironment's
getWorker the first time a javascript/typescript worker is requested —
the earliest point a user actually needs node typings. Monaco
propagates extra libs added after worker creation, so completions
pick the typings up as soon as the chunk lands.

Entry chunk: 2,648 KB → 238 KB (gzip 518 KB → 74 KB). The nodeTypes
chunk (2,410 KB) now loads only on first JS/TS editor use.

Note: the editor.api chunk itself is still modulepreloaded because app
code (KsEditor, language configurators) imports it statically — that
is a separate, larger refactor.
