+++ ui/src/main.ts
@@ -3,20 +3,6 @@ import type {Router} from "vue-router"
 
 import "./utils/monacoEnvironment"
 
-const NodeTypesRaw = import.meta.glob("/node_modules/@types/node/**/*.d.ts", {eager: true, query: "?raw", import: "default"}) as Record<string, string>
-function loadNodeTypes(tries = 0) {
-    import("monaco-editor/esm/vs/editor/editor.api").then(({languages}) => {
-        if (languages.typescript) {
-            for (const path in NodeTypesRaw) {
-                languages.typescript.typescriptDefaults.addExtraLib(NodeTypesRaw[path], `file://${path}`)
-            }
-        } else if (tries <= 15) {
-            setTimeout(() => loadNodeTypes(tries + 1), (tries + 1) * 100)
-        }
-    })
-}
-loadNodeTypes()
-
 import App from "./App.vue"
 import initApp from "./utils/init"
 import {setupKestraHttp} from "./utils/kestraHttp"
